Move legacy controllers and transformers into a legacy namespace.
